Reestruturação do repositorio, código multiplataforma, bugs corrigidos e novas funcionalidades
O repositório foi reestruturado para conter o código completo das suas
dependencias e tentar reduzir o tamanho geral.
O código agora suporta as plataformas Windows, Mac OS X, Android e iOS,
mas somente as bibliotecas para windows estão compiladas e possuem um
projeto configurado, bem como um projeto de base.
Alguns bugs foram corrigidos e novas funcionalidades foram adicionadas.

Bugs Corrigidos:
- Corrigido bug na classe Som, que corrompia a heap quando um som era
destruido antes de terminar de tocar.
- Corrigido bug que calculava a posição errada do toque em tela cheia.
- Corrigido bug que fazia o controle de tempo oscilar demasiadamente. O
atual controle de tempo é altamente estavel.

Adicionado:
- Adicionado suporte ao vibrador do joystick.
- Adicionado método na classe Texto para ajustar a string em uma largura
especificada.

Modificado:
- Reestruturação do repositorio.
- Ajustes no código para suportar multiplas plataformas.

Removido:
- Removido variaveis não utilizadas em alguns métodos.
- Removidos arquivos desnecessarios.
- Removidos projetos de exemplo.
- Removidos projetos do MSVS 2012
- Removido header que fazia a configuração automática para projetos no
Visual Studio (uniConfig.h)

+++ libUnicornio/src/TileMap/CamadaDeObjetos.cpp
@@ -145,7 +145,7 @@ bool CamadaDeObjetos::existeObjetoDoTipoNaPos(string tipo, float tx, float ty)
 
 int CamadaDeObjetos::getNumObjetos()
 {
-	return objetos.size();
+	return (int)objetos.size();
 }
 
 ObjetoTile *CamadaDeObjetos::getObjeto(int indice)
